Moved functions out of minkasi_jax.py
